refactored Trip component
